dayflow-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic<2.0
sqlalchemy
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv
email-validator
